+++ Lab5/src/lab05/Tipo.java
@@ -4,5 +4,5 @@ public abstract class Tipo {
 	public abstract String toString();
 	public abstract int getId();
 	public abstract double getValor();
-	public int valor;
+	public double valor;
 }
